+++ cmd/gateway/main.go
@@ -42,7 +42,7 @@ import (
 var (
 	configFile     string
 	enableTelegram bool
-	Version        = "1.17.0-rc2"
+	Version        = "1.18.0-rc1"
 	GitCommit      = "dev" // 编译时通过 ldflags 注入，如未注入则显示 dev
 )